+++ src/components/GenericPagesComponents/Select/Select.tsx
@@ -5,15 +5,13 @@ import { useObjectRef } from "@react-aria/utils";
 import { useSelectState } from "react-stately";
 import { useSelect, useButton, mergeProps, useFocusRing } from "react-aria";
 
-import UnstyledButton from "../UnstyledButton";
+import UnstyledButton from "@/components/UnstyledButton";
 import {
   InputFocusUnderline,
   RotatedInputLabel,
-} from "../SharedComponents/Input/Input";
-
-import { ListBox } from "./ListBox";
-import { Popover } from "./Popover";
-
+} from "@/components/SharedComponents/Input/Input";
+import { ListBox } from "@/components/SharedComponents/ListBox";
+import { Popover } from "@/components/SharedComponents/Popover";
 import getColorByName from "@/styles/themeHelpers/getColorByName";
 import ellipsis from "@/styles/ellipsis";
 import getColorByLocation from "@/styles/themeHelpers/getColorByLocation";
