+++ src/components/GenericPagesComponents/Select/index.tsx
@@ -1,0 +1,2 @@
+export { Select, Item } from "./Select";
+export type { SelectItem } from "./Select";
