+++ src/components/SharedComponents/Popover/index.tsx
@@ -1,0 +1,1 @@
+export { Propover } from "./Popover";
